Updated location for PAF adapter
